Create components folder, move App component, fix tests
